fix: ArtifactVersion.CreateTime as int64 instead of float64

+++ artifact/service.go
@@ -267,7 +267,7 @@ type ArtifactVersion struct {
 	Version        int64
 	CanonicalURI   string
 	CustomMetadata map[string]any
-	CreateTime     float64
+	CreateTime     int64
 	MimeType       string
 }
